Endpoint example: 	modified:   backend/src/earnings_estimate/earnings_estimate_controller.py 	modified:   backend/src/earnings_estimate/earnings_estimate_manager.py 	modified:   backend/src/main.py 	new file:   backend/src/price/__init__.py 	modified:   backend/src/price/price_controller.py 	modified:   docker-compose.yml
