+++ backend/src/main.py
@@ -1,23 +1,10 @@
 from fastapi import FastAPI
-from earnings_estimate.earnings_estimate_controller import earnings_estimate
+
+from .earnings_estimate import earnings_estimate_controller
+from .price import price_controller
 
 app = FastAPI()
 
 
-@app.get("/")
-async def root():
-    return {"message": "Hello World"}
-
-@app.get("/retrieve/price/{symbol}")
-async def price(symbol : str):
-
-
-
-    return {"message : "+symbol}
-
-@app.get("/retrieve/earnings_estimate/{symbol}")
-async def price(symbol : str):
-
-    get_earnings_estimate = earnings_estimate.get_earnings_estimate(symbol)
-
-    return get_earnings_estimate
+app.include_router(earnings_estimate_controller.router)
+app.include_router(price_controller.router)
